Add files via upload
so basically i added test to transactions, added tests to catalogger and added test and created the goal, and goaltagtracker doas and entities. The main problem right now is that deleleting a goal is broken I think because of how my goal entity is set up. If you take a look I am sure youll spot my mistake in a second.